+++ SFExt/config/config.js
@@ -11,6 +11,7 @@ function save_options() {
   let qnotifyweb = document.getElementById('qnotifyweb').checked;
   let webhook = document.getElementById('webhook').value;
   let refEmail = document.getElementById('refemail').value;
+  let signature = document.getElementById('signature').value;
   let customurls = document.getElementById('customurls').value;
   let grabLink = document.getElementById('grabLink').checked;
   let wideCase = document.getElementById('wideCase').checked;
@@ -24,6 +25,7 @@ function save_options() {
       savedQNotifyWeb: qnotifyweb,
       savedWebhook: webhook,
       savedRefEmail: refEmail,
+      savedSig: signature,
       savedURLS: customurls,
       savedGrab: grabLink,
       savedWide: wideCase
@@ -37,7 +39,7 @@ function save_options() {
 }
 
 function reset_options() {
-  chrome.storage.sync.remove(["savedTimeout", "savedDefect", "savedPP", "savedEDU", "savedQueue", "savedQNotify", "savedQNotifyWeb", "savedWebhook", "savedRefEmail", "savedURLS", "savedGrab", "savedWide"], function() {
+  chrome.storage.sync.remove(["savedTimeout", "savedDefect", "savedPP", "savedEDU", "savedQueue", "savedQNotify", "savedQNotifyWeb", "savedWebhook", "savedRefEmail", "savedSig", "savedURLS", "savedGrab", "savedWide"], function() {
       let error = chrome.runtime.lastError;
       if (error) {
           console.error(error);
@@ -57,6 +59,7 @@ function restore_options() {
       savedQNotifyWeb: false,
       savedWebhook: '',
       savedRefEmail: '',
+      savedSig: '',
       savedURLS: `{"SFExt":"${configURL}"}`,
       savedUUID: '',
       savedGrab: false,
@@ -71,6 +74,7 @@ function restore_options() {
       document.getElementById('qnotifyweb').checked = result.savedQNotifyWeb;
       document.getElementById('webhook').value = result.savedWebhook;
       document.getElementById('refemail').value = result.savedRefEmail;
+      document.getElementById('signature').value = result.savedSig;
       document.getElementById('customurls').value = result.savedURLS;
       document.getElementById('grabLink').checked = result.savedGrab;
       document.getElementById('wideCase').checked = result.savedWide;
@@ -95,6 +99,7 @@ function export_options() {
         savedQNotifyWeb: false,
         savedWebhook: '',
         savedRefEmail: '',
+        savedSig: '',
         savedURLS: `{"SFExt":"${configURL}"}`,
         savedUUID: '',
         savedGrab: false,
@@ -146,6 +151,7 @@ function import_options() {
             savedQNotifyWeb: json.savedQNotifyWeb,
             savedWebhook: json.savedWebhook,
             savedRefEmail: json.savedRefEmail,
+            savedSig: json.savedSig,
             savedURLS: json.savedURLS,
             savedUUID: json.savedUUID,
             savedGrab: json.savedGrab,